requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]
python-multipart

# Asynchronous Task Queue
celery
redis

# Data Parsing & OCR
pandas
openpyxl
pdfplumber
pytesseract
Pillow

# LLM Client
google-generativeai

# Configuration & Utilities
pydantic-settings
python-dotenv
jsonschema

# Authentication & Database
SQLAlchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
